Calculate curve shape over an specific region

calc_curve_shape() calculates ROC curve shape over a prespecified region. The prespecified region can be either a range of values in $TPR$ axis, or in $FPR$ axis.

Usage

calc_curve_shape(
  data = NULL,
  response = NULL,
  predictor = NULL,
  lower_threshold,
  upper_threshold,
  ratio
)

Arguments

data

A data.frame or extension (e.g. a tibble) containing values for predictors and response variables.

response

A data variable which must be a factor, integer or character vector representing the class to predict on each observation or Gold Standard. For more info on how to select class of interest see Methods below.

predictor

A data variable wich must be numeric, representing values of a classifier or predictor for each observation.

lower_threshold, upper_threshold

Ranges in which to calculate curve shape, they can take values from 0 to 1.

ratio

If "tpr", shape will calculated over TPR ratio. Otherwise, if "fpr" it will be calculated over FPR ratio.

Value

A character indicating ROC curve shape in the specified region. Result can take any of the following values:

Concave. ROC curve is concave over the entire region.

Partially proper. ROC curve loses concavity at some point of the region.

Hook under chance. ROC curve loses concavity at some point of the region and it lies below chance line.

Data masking variables

Both response and predictor support data-masking variables. For more information please check General usage in this article.

Nevertheless, function may use environment variables as well. In order to use this type of variables simply set data = NULL and provide response and predictor as normal, whithout data-masking.

Examples

# Calc curve shape in TPR region (0.9, 1)
calc_curve_shape(iris, Species, Sepal.Width, 0.9, 1, "tpr")
#> Warning: [*] Upper threshold already included in points ...
#> [1] "Partially proper"