Add fpauc lower bound to a ROC plot

Calculate and plot lower bound defined by FpAUC sensitivity index.

Usage

add_fpauc_partially_proper_lower_bound(
  data = NULL,
  fpr = NULL,
  tpr = NULL,
  response = NULL,
  predictor = NULL,
  threshold
)

add_fpauc_concave_lower_bound(
  data = NULL,
  fpr = NULL,
  tpr = NULL,
  response = NULL,
  predictor = NULL,
  threshold
)

add_fpauc_lower_bound(
  data = NULL,
  fpr = NULL,
  tpr = NULL,
  response = NULL,
  predictor = NULL,
  threshold
)

Arguments

data

A data.frame or extension (e.g. a tibble) containing values for predictors and response variables.

fpr, tpr

Numeric vectors representing FPR and TPR for the classifier. These arguments will be used when no data, response and predictor are supplied.

response

A data variable which must be a factor, integer or character vector representing the class to predict on each observation or Gold Standard. For more info on how to select class of interest see Methods below.

predictor

A data variable wich must be numeric, representing values of a classifier or predictor for each observation.

threshold

A number between 0 and 1, representing lower TPR for the region of interest.

Details

add_fpauc_lower_bound() provides a upper level function which automatically calculates curve shape and plots a lower bound that ajust to it.

To enforce specific bounds lower level functions add_fpauc_partially_proper_lower_bound and add_fpauc_concave_lower_bound are provided, which plot lower bound when curve shape partially proper ( presents some kind of hook) or is concave in region of interest.

Examples

# Add lower bound based on cuve shape (Concave)
plot_roc_curve(iris, response = Species, predictor = Sepal.Width) +
  add_fpauc_lower_bound(
    data = iris,
    response = Species,
    predictor = Sepal.Width,
    threshold = 0.9
  )
#> Error: argument ".condition" is missing, with no default