Sensitivity indexes

Sensitivity indexes provide different ways of calculating partial area under ROC curve in a specific TPR range. Two different approaches to calculate this area are available:

fp_aup() applies fitted partial area under curve index (FpAUC). This one calculates area under curve adjusting to points defined by curve in selected region.

np_auc() applies normalized partial area under curve index (NpAUC), which calculates area under curve over the whole specified region.

Usage

fp_auc(data = NULL, response, predictor, lower_tpr)

np_auc(data = NULL, response, predictor, lower_tpr)

Arguments

data

A data.frame or extension (e.g. a tibble) containing values for predictors and response variables.

response

A data variable which must be a factor, integer or character vector representing the class to predict on each observation or Gold Standard. For more info on how to select class of interest see Methods below.

predictor

A data variable wich must be numeric, representing values of a classifier or predictor for each observation.

lower_tpr

Lower value of TPR for the region in which to calculate the partial area under curve.

Value

A numeric indicating the partial area under curve in the selected region.

Methods

When performing calculations response will be transformed internally to a factor with levels (0,1), which represent absence or presence of the class we want to predict. In order to select which class is the one to predict among availables in response, function follows some criteria according to var type:

integer. When working with an integer vector, function will consider the smallest as the class to predict.

character. When working with a character vector, function will consider the first value after using sort() over all posible options.

factor. When working with a factor var, function will select first class in levels().

Al other classes which are not identified as the class to predict will be collapsed into 0. For this reason, try to provide convenient names when working or alternatively provide a factor where levels are equal to 0 and 1, being 1 cases with the condition of interest.

Data masking variables

Both response and predictor support data-masking variables. For more information please check General usage in this article.

Nevertheless, function may use environment variables as well. In order to use this type of variables simply set data = NULL and provide response and predictor as normal, whithout data-masking.

References

Franco M. y Vivo J.-M. Evaluating the Performances of Biomarkers over a Restricted Domain of High Sensitivity. Mathematics 9, 2826 (2021).

Jiang Y., Metz C. E. y Nishikawa R. M. A receiver operating characteristic partial area index for highly sensitive diagnostic tests. Radiology 201, 745-750 (1996).

Examples

# Calculate fp_auc on TPR region (0.9, 1)
fp_auc(iris, response = Species, predictor = Sepal.Width, lower_tpr = 0.9)
#> Warning: [*] Upper threshold already included in points ...
#> [1] 0.8516667
# Calculate np_auc on TPR region (0.9, 1)
np_auc(iris, response = Species, predictor = Sepal.Width, lower_tpr = 0.9)
#> Warning: [*] Upper threshold already included in points ...
#> [1] 0.472