Specificity indexes

Specifity indexes provide different ways of calculating partial area under ROC curve in a specific FPR range. Two different approaches to calculate this area are available:

sp_aup() applies standardized partial area under curve index (SpAUC). This one calculates area under curve adjusting to points defined by curve in selected region.

tp_auc() applies tighter partial area under curve index (TpAUC), which calculates area under curve over the whole specified region.

Usage

sp_auc(data = NULL, response, predictor, lower_fpr, upper_fpr)

tp_auc(data = NULL, response, predictor, lower_fpr, upper_fpr)

Arguments

data

A data.frame or extension (e.g. a tibble) containing values for predictors and response variables.

response

A data variable which must be a factor, integer or character vector representing the class to predict on each observation or Gold Standard. For more info on how to select class of interest see Methods below.

predictor

A data variable wich must be numeric, representing values of a classifier or predictor for each observation.

lower_fpr, upper_fpr

Lower and upper values of FPR for the region in which to calculate the partial area under curve.

Value

A numeric indicating the partial area under curve in the selected region.

Methods

When performing calculations response will be transformed internally to a factor with levels (0,1), which represent absence or presence of the class we want to predict. In order to select which class is the one to predict among availables in response, function follows some criteria according to var type:

integer. When working with an integer vector, function will consider the smallest as the class to predict.

character. When working with a character vector, function will consider the first value after using sort() over all posible options.

factor. When working with a factor var, function will select first class in levels().

Al other classes which are not identified as the class to predict will be collapsed into 0. For this reason, try to provide convenient names when working or alternatively provide a factor where levels are equal to 0 and 1, being 1 cases with the condition of interest.

Data masking variables

Both response and predictor support data-masking variables. For more information please check General usage in this article.

Nevertheless, function may use environment variables as well. In order to use this type of variables simply set data = NULL and provide response and predictor as normal, whithout data-masking.

References

McClish D. K. Analyzing a Portion of the ROC Curve. Medical Decision Making 9, 190-195 (1989).

Vivo J.-M., Franco M. y Vicari D. Rethinking an ROC partial area index for evaluating the classification performance at a high specificity range. Advances in Data Analysis and Classification 12, 683-704 (2018).

Examples

# Calc sp_auc in FPR region (0, 0.1)
sp_auc(
 iris,
 response = Species,
 predictor = Sepal.Width,
 lower_fpr = 0,
 upper_fpr = 0.1
)
#> Corrected partial area under the curve (specificity 1-0.9): 0.7242
# Calc tp_auc in FPR region (0, 0.1)
tp_auc(
 iris,
 response = Species,
 predictor = Sepal.Width,
 lower_fpr = 0,
 upper_fpr = 0.1
)
#> Warning: [*] Lower threshold already included in points ...
#> Warning: [*] Upper threshold already included in points ...
#> [1] 0.7212121